requirement.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn
seaborn
catboost
xgboost
flask 
drill
mlflow
#-e .